minor changes in Adhesion Factor drafts

+++ Source/Drafts/Adhesion_Factor/Adhesion_Factor.h
@@ -10,12 +10,12 @@ struct Adhesion_Factor
     // This packet is used when the trackside requests a change of the adhesion factor to be used in the brake model.
     // Packet Number = 71
 
-    uint8_t   Q_DIR;            // # 2
+    uint32_t  Q_DIR;            // # 2
     uint16_t  L_PACKET;         // # 13
-    uint8_t   Q_SCALE;          // # 2
+    uint32_t  Q_SCALE;          // # 2
     uint16_t  D_ADHESION;       // # 15
     uint16_t  L_ADHESION;       // # 15
-    uint8_t   M_ADHESION;       // # 1
+    uint32_t  M_ADHESION;       // # 1
 };
 
 typedef struct Adhesion_Factor Adhesion_Factor;
